web-backend/homeAutomationApi/requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.0
charset-normalizer==2.1.0
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.13
django-cors-headers==3.13.0
django-phonenumber-field==6.3.0
django-restframework==0.0.1
django-twilio==0.13.2
django-urls==1.1.3
django-webpush==0.3.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-extra-fields==3.4.0
hiredis==2.0.0
http-ece==1.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.4
phonenumbers==8.12.50
Pillow==9.1.1
psycopg==3.0.15
psycopg2==2.9.3
py-vapid==1.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==22.0.0
python-decouple==3.6
pytz==2022.1
pywebpush==1.9.4
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
twilio==6.63.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
zope.interface==5.4.0
